Close the async read thread once task finishes (#66)
A thread is taking resources, e.g. memory and cpu cores. So this PR
tries to close the async read thread after the task is done.

Shutting down a Java ExecutorService here is likely to be lightweight
and not blocking for normal cases, since all the batches should be read
and consumed when a task completes.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>