Fix pom file of module `rapids-velox` (#110)
Add the missing `provided` tag for `spark-sql-columnar-shims-spark32` in
case that it is loaded for multiple times.

The pom file of rapids-velox was directly copied from the 24.X version,
in which the `provided` tag for `spark-sql-columnar-shims-spark32` was
defined at the parent pom file. However, the refactor for 25.04
streamlined the parent pom file and removed all dependencies of gluten.

Signed-off-by: sperlingxx <lovedreamf@gmail.com>