No task context in async thread to avoid it becoming dedicated thread (#182)
![image](https://github.com/user-attachments/assets/17226293-5ad4-401e-a0aa-f3901348a133)

If a backend thead meets two conditions:
1.  TaskContext.get() != null
2.  once called retry iterator
Then it will be registered as a dedicated thread for a task.

If the shuffle read async thread becomes a dedicated thread, it will
also cause premature concensue of OOM state machine, just like
https://github.com/NVIDIA/spark-rapids/issues/12311

<!--

1. PR description

4. Configs for new feature if any(no **sensitive** content please)

5. Dorado ids to test for this change

6. PIC(who should help you do the test at BD site)

Please also assign the PIC as the assignee, so he will receive this test
command.

-->

Signed-off-by: Hongbin Ma (Mahone) <mahongbin@apache.org>