VeloxScan: Fix the computation of the pre-allocation size of leaf buffers (#235)
The mirror PR of
https://github.com/HighPerfDataAccelerator/spark-rapids/pull/233 which
is back ported to 25.06.

Previously, the `rapids-hybrid-execution` module reworked the
computation logic about pre-allocation size of leaf buffers: no longer
estimated the pre-allocation size by the last batch, instead estimated
based on the accumulated size of all processed batches.

However, this commit also introduced a bug: when there are enormous
batches being processed, the accumulated `dataSize` or `offsetSize` may
overflow the INT_32, should have used INT_64.

This PR is to fix this bug which will be triggered under extreme large
data scenario. And this PR does not refer any PR of the public repo,
since this PR is copied from the `rapids-hybrid-execution` repo.

Signed-off-by: sperlingxx <lovedreamf@gmail.com>